Move relevant FormikComponents to articlepage
